System.IO.LockFile.Internal

Description

Low-level API for providing exclusive access to a resource using lock file.

Locking primitives

lock :: (MonadMask m, MonadIO m) => LockingParameters -> FilePath -> m Handle

Open lock file write PID of a current process in to it and return its handle.

If operation doesn't succeed, then LockingException is raised. See also LockingParameters and RetryStrategy for details.

unlock :: (MonadMask m, MonadIO m) => FilePath -> Handle -> m ()

Close lock file handle and then delete it.

Configuration

data LockingParameters

Locking algorithm parameters. When in doubt, use def, otherwise start with it. Example:

lockedDo
    :: (MonadMask m, MonadIO m)
    => FilePath
    -> m a
    -> m a
lockedDo = withLockFile lockParams lockFile
  where
    lockParams = def
        { retryToAcquireLock = NumberOfTimes 3
        }

    lockFile = withLockExt "/var/lock/my-app"

Constructors

LockingParameters

Fields

retryToAcquireLock :: !RetryStrategy

Strategy for handling situations when lock-file is already acquired.

sleepBetweenRetries :: !Word64

Sleep interval in microseconds.

data RetryStrategy

Defines strategy for handling situations when lock-file is already acquired.

Constructors

No
Don't retry at all.

Indefinitely
Retry indefinitely.

NumberOfTimes !Word8
Retry only specified number of times. If equal to zero then it is interpreted same way as No.

Exceptions

data LockingException

Constructors

UnableToAcquireLockFile FilePath
Wasn't able to aquire lock file specified as an argument.

CaughtIOException IOException
IOException occurred while creating or removing lock file.